requirements.txt
amqp==2.5.2
asgiref==3.2.7
astroid==2.3.3
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
Django==3.0.4
django-celery-beat==2.0.0
django-phonenumber-field==4.0.0
django-simple-history==2.8.0
django-timezone-field==4.0
importlib-metadata==1.6.0
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.0
psycopg2-binary==2.8.4
pylint==2.4.4
python-crontab==2.4.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
redis==3.4.1
sentry-sdk==0.14.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
zipp==3.1.0
